input password with hide/show button
